+++ libs/wire-api/default.nix
@@ -51,6 +51,7 @@
 , iproute
 , iso3166-country-codes
 , iso639
+, jose
 , lens
 , lib
 , memory
@@ -152,6 +153,7 @@ mkDerivation {
     iproute
     iso3166-country-codes
     iso639
+    jose
     lens
     memory
     metrics-wai
